refactor: Study -> Schedule 패키지로 수정
