+++ src/main/java/com/gobookee/schedule/controller/SearchStudyCountAjaxHandlerServlet.java
@@ -1,7 +1,7 @@
-package com.gobookee.study.controller;
+package com.gobookee.schedule.controller;
 
 import com.gobookee.common.JsonConvertTemplate;
-import com.gobookee.study.service.StudyService;
+import com.gobookee.schedule.service.ScheduleService;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -11,16 +11,16 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.sql.Date;
 
-@WebServlet(name = "searchStudyCountAjaxHandlerServlet", urlPatterns = "/study/searchstudycount")
+@WebServlet(name = "searchStudyCountAjaxHandlerServlet", urlPatterns = "/schedule/searchstudycount")
 public class SearchStudyCountAjaxHandlerServlet extends HttpServlet {
-    private StudyService studyService = StudyService.studyService();
+    private ScheduleService scheduleService = ScheduleService.scheduleService();
 
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         String stringDate = request.getParameter("date");
         Long placeSeq = Long.valueOf(request.getParameter("placeSeq"));
         Date date = Date.valueOf(stringDate);
-        int count = studyService.getStudyCountByPlaceSeqAndDate(placeSeq, date);
+        int count = scheduleService.getStudyCountByPlaceSeqAndDate(placeSeq, date);
         JsonConvertTemplate.toJson(count, response);
     }
