+++ src/main/java/com/gobookee/schedule/service/ScheduleService.java
@@ -3,6 +3,7 @@ package com.gobookee.schedule.service;
 import com.gobookee.schedule.model.dao.ScheduleDao;
 
 import java.sql.Connection;
+import java.sql.Date;
 import java.util.HashMap;
 
 import static com.gobookee.common.JDBCTemplate.*;
@@ -31,9 +32,16 @@ public class ScheduleService {
             close(conn);
             return true;
         }
-
         rollback(conn);
         close(conn);
         return false;
     }
+
+    public int getStudyCountByPlaceSeqAndDate(Long placeSeq, Date date) {
+        conn = getConnection();
+        int count = scheduleDao.getStudyCountByPlaceSeqAndDate(conn, placeSeq, date);
+        close(conn);
+        return count;
+    }
+
 }
